implements and test basic download process

+++ chazutsu/datasets/framework/dataset.py
@@ -1,6 +1,7 @@
 import os
 import re
-from random import random
+import mmap
+import random
 import zipfile
 import tarfile
 import requests
@@ -41,7 +42,9 @@ class Dataset():
         #   test_size: have to be smaller than 1
         self.check_directory(directory)
         dataset_root = os.path.join(directory, self.name)
-        
+        if not os.path.isdir(dataset_root):
+            os.mkdir(dataset_root)
+
         # download and save file
         save_file_path = self.save_dataset(dataset_root)
 
@@ -49,16 +52,17 @@ class Dataset():
         extracted_file_path = self.extract(save_file_path)
 
         # split to train & test
-        self.train_test_split(extracted_file_path, test_size)
+        train_test_path = self.train_test_split(extracted_file_path, test_size)
     
         # make sample file
-        self.make_samples(save_file_path, sample_size)
+        sample_path = self.make_samples(extracted_file_path, sample_size)
         
         # remove raw file
-        if not keep_raw:
+        if not keep_raw and os.path.isfile(extracted_file_path):
             os.remove(extracted_file_path)
         
         self.logger.info("Done all process!")
+        return dataset_root
     
     def save_dataset(self, dataset_root):
         # download and save it as raw file
@@ -115,6 +119,7 @@ class Dataset():
                         extracteds.append(p)
         
         if remove:
+            # remove downloaded raw file (zip/tar.gz etc)
             os.remove(path)
 
         return extracteds
@@ -122,53 +127,66 @@ class Dataset():
     def train_test_split(self, original_file_path, test_size):
         if test_size < 0 or test_size > 1:
             self.logger.error("test_size have to be between 0 ~ 1. if you don't want to split, please set 0.")
-            return None
+            return []
         elif test_size == 0 or test_size == 1:
-            return None
+            return []
 
         self.logger.info("Split to train & test file.")
 
         total_count = self.get_line_count(original_file_path)
-        def is_test():
-            r = random.random()
-            return True if r <= test_size else False
+        test_count = int(round(total_count * test_size))
+        test_targets = random.sample(range(total_count), test_count)
         
         base, ext = os.path.splitext(original_file_path)
-        train_file = open(base + "_train" + ext, "wb")
-        test_file = open(base + "_test" + ext, "wb")
+        train_test_path = [base + x + ext for x in ["_train", "_test"]]
+        train_file = open(train_test_path[0], "wb")
+        test_file = open(train_test_path[1], "wb")
 
-        test_count = 0
         with open(original_file_path, "rb") as f:
+            i = 0
             for line in tqdm(f, total=total_count):
-                target, tc = test_file, 1 if is_test() else train_file, 0
+                target = test_file if i in test_targets else train_file
                 target.write(line)
-                test_count += tc
+                i += 1
 
         train_file.close()
         test_file.close()
 
-        self.logger.info("Splited to {} train & {} test({:.2f}%).".format(
+        to_base = lambda x: os.path.basename(x)
+        self.logger.info("File is splited to {} & {}. Each records are {} & {} (test_size={:.2f}%).".format(
+            to_base(train_test_path[0]), to_base(train_test_path[1]),
             total_count - test_count, test_count, test_count / total_count * 100)
             )
-        return None
+        return train_test_path
 
-    def make_samples(self, original_file_path, sample_size):
-        if sample_size == 0:
-            return None
+    def make_samples(self, original_file_path, sample_count):
+        if sample_count == 0:
+            return ""
         
-        self.logger.info("Make sample file.")
         base, ext = os.path.splitext(original_file_path)
-        samples_file = open(base + "_samples" + ext, "wb")
+        sample_path = base + "_samples" + ext
+        samples_file = open(sample_path, "wb")
+
+        total_count = self.get_line_count(original_file_path)
+        # for reproducibility of sampling, use fixed interval
+        sample_target = range(0, total_count, round(total_count / sample_count))
         with open(original_file_path, "rb") as f:
             count = 0
+            i = 0
             for line in f:
-                samples_file.write(line)
-                count += 1
-                if count > sample_size:
+                if i in sample_target:
+                    samples_file.write(line)
+                    count += 1
+                i += 1
+                if count == sample_count:
                     break
-        sample_file.close()
 
-        return None
+        samples_file.close()
+        self.logger.info("Make {} by picking {} records from original file.".format(
+            os.path.basename(sample_path), sample_count)
+        )
+
+        return sample_path
 
     def _get_file_name(self, resp):
         cd = resp.headers["content-disposition"]
@@ -180,7 +198,7 @@ class Dataset():
 
         return file_name
 
-    def get_line_count(file_path):  
+    def get_line_count(self, file_path):  
         count = 0
         with open(file_path, "r+") as fp:
             buf = mmap.mmap(fp.fileno(), 0)
